remove filter state from daoadmin context
